fix: 修复循环依赖和重试逻辑问题
1. 修复循环依赖问题
   - 创建独立的 relay/cache 包
   - 将 HandleClaudeCache 从 relay/controller 移到 relay/cache
   - 解除 relay/channel/aws/claude 和 relay/controller 之间的循环依赖

2. 修复 CacheGetRandomSatisfiedChannel 参数不匹配
   - 添加缺失的 responseID 参数（支持 Claude 缓存定向）
   - 添加 excludeChannelIds 参数（排除已失败渠道）

3. 改进重试逻辑
   - 使用 failedChannelIds 列表记录所有失败渠道
   - 避免重复选择已失败的渠道
   - 支持通过 X-Response-ID 进行 Claude 缓存定向

修改的文件：
- relay/cache/claude.go (新建)
- relay/channel/aws/claude/main.go
- relay/controller/claude.go
- controller/relay.go
